# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Battle Rap API
  version: 1.0.0
  description: >
    REST API for managing online rap battles, tournaments, submissions, judging,
    and administration workflows.
servers:
  - url: https://api.example.com
    description: Production
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Registration succeeded.
          headers:
            Set-Cookie:
              description: Refresh token cookie (`refresh_token`).
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthWithUserResponse'
        '409':
          $ref: '#/components/responses/Conflict'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login succeeded.
          headers:
            Set-Cookie:
              description: Refresh token cookie (`refresh_token`).
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthWithUserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      operationId: refreshToken
      security:
        - RefreshCookie: []
      responses:
        '200':
          description: Access token refreshed.
          headers:
            Set-Cookie:
              description: Refreshed `refresh_token` cookie.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: Logout and revoke refresh token
      operationId: logout
      security:
        - RefreshCookie: []
      responses:
        '204':
          description: Logged out successfully.
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: Get authenticated user profile snapshot
      operationId: getAuthMe
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current user profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileView'
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
  /api/v1/profile/me:
    get:
      tags: [Profile]
      summary: Get current user profile
      operationId: getOwnProfile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Authenticated user profile view.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileView'
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
  /api/v1/admin/roles/{userId}:
    post:
      tags: [Admin]
      summary: Set user role
      operationId: setUserRole
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleChangeRequest'
      responses:
        '200':
          description: Updated user roles.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRolesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/admin/finalize/battles/{id}:
    post:
      tags: [Admin]
      summary: Finalize battle and refresh views
      operationId: finalizeBattle
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Battle finalized.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/admin/users:
    get:
      tags: [Admin]
      summary: List users with roles
      operationId: listUsersForAdmin
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - in: query
          name: search
          schema:
            type: string
            description: Case-insensitive match against display name, email, or full name.
        - in: query
          name: role
          schema:
            type: string
            enum: [admin, moderator, artist, judge, listener]
        - in: query
          name: sort
          schema:
            type: string
            enum: [created_at, -created_at, display_name, -display_name]
      responses:
        '200':
          description: Paginated user list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/admin/users/{id}:
    get:
      tags: [Admin]
      summary: Get user profile as admin
      operationId: getUserProfileAdmin
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Full user profile with private fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileView'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/admin/audit-log:
    get:
      tags: [Admin]
      summary: List audit log entries
      operationId: listAuditLog
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - in: query
          name: actor_id
          schema:
            type: string
            format: uuid
        - in: query
          name: action
          schema:
            type: string
        - in: query
          name: target_table
          schema:
            type: string
        - in: query
          name: target_id
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Audit log entries.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Access token issued by the API. Required for authenticated endpoints.
    RefreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: >
        Refresh token cookie issued during login/registration.
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: Authentication required or token invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Insufficient permissions.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: Resource conflict.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    ValidationError:
      description: Validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        detail:
          type: string
        status:
          type: integer
        field_errors:
          type: object
          additionalProperties:
            type: string
      example:
        type: https://httpstatuses.com/422
        title: Payload validation failed.
        detail: Payload validation failed.
        status: 422
        field_errors:
          email: Invalid email.
...
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - display_name
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        display_name:
          type: string
          minLength: 2
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    AuthUser:
      type: object
      required:
        - id
        - email
        - display_name
        - roles
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        display_name:
          type: string
        roles:
          type: array
          items:
            type: string
    AuthWithUserResponse:
      type: object
      required:
        - access_token
        - token_type
        - expires_in
        - user
      properties:
        access_token:
          type: string
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          description: Access token TTL in seconds.
        user:
          $ref: '#/components/schemas/AuthUser'
    ProfileAvatar:
      type: object
      required:
        - key
        - url
      properties:
        key:
          type: string
          description: Internal storage key for the avatar asset.
        url:
          type: string
          format: uri
          description: CDN URL for the avatar asset.
    ProfileViewerContext:
      type: object
      required:
        - is_self
        - can_edit
        - can_moderate
        - can_view_private
      properties:
        is_self:
          type: boolean
          description: Indicates the profile belongs to the caller.
        can_edit:
          type: boolean
          description: Caller may submit profile change requests.
        can_moderate:
          type: boolean
          description: Caller has moderator or admin privileges over this profile.
        can_view_private:
          type: boolean
          description: Caller may see private fields such as email or full name.
    ProfileView:
      type: object
      required:
        - id
        - display_name
        - roles
        - created_at
        - updated_at
        - viewer_context
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: string
        roles:
          type: array
          items:
            type: string
        avatar:
          $ref: '#/components/schemas/ProfileAvatar'
          nullable: true
        bio:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        viewer_context:
          $ref: '#/components/schemas/ProfileViewerContext'
        email:
          type: string
          format: email
          description: Present when the caller can view private information.
        age:
          type: integer
          minimum: 12
          maximum: 120
          nullable: true
        vk_id:
          type: string
          nullable: true
        full_name:
          type: string
          nullable: true
        socials:
          type: object
          additionalProperties: true
          description: Social handles map, available to privileged viewers.
...
    AdminUser:
      type: object
      required:
        - id
        - email
        - display_name
        - roles
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        display_name:
          type: string
        roles:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        last_login_at:
          type: string
          format: date-time
          nullable: true
    AdminUserListResponse:
      type: object
      required:
        - data
        - page
        - limit
        - total
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AdminUser'
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
    AuditLogEntry:
      type: object
      required:
        - id
        - action
        - target_table
        - created_at
      properties:
        id:
          type: string
          format: uuid
        actor_user_id:
          type: string
          format: uuid
          nullable: true
        actor_display_name:
          type: string
          nullable: true
        action:
          type: string
        target_table:
          type: string
        target_id:
          type: string
          format: uuid
          nullable: true
        payload:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
    AuditLogResponse:
      type: object
      required:
        - data
        - page
        - limit
        - total
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AuditLogEntry'
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
    RefreshResponse:
      type: object
      required:
        - access_token
        - token_type
        - expires_in
      properties:
        access_token:
          type: string
        token_type:
          type: string
        expires_in:
          type: integer
...
    RoleChangeRequest:
      type: object
      required:
        - op
        - role
      properties:
        op:
          type: string
          enum: [grant, revoke]
        role:
          type: string
          enum: [artist, judge, listener, moderator, admin]
    UserRolesResponse:
      type: object
      required:
        - user_id
        - roles
      properties:
        user_id:
          type: string
          format: uuid
        roles:
          type: array
          items:
            type: string
...